server/requirements.txt
alembic==1.12.1
aniso8601==9.0.1
bcrypt==4.0.1
blinker==1.6.3
click==8.1.7
Faker==19.12.1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
gunicorn==21.2.0
honcho==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
load-dotenv==0.1.0
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.22
SQLAlchemy-serializer==1.4.1
typing_extensions==4.8.0
Werkzeug==3.0.1
WTForms==3.1.0
